pyproject.toml
[project]
name = "gainstracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=24.10.0",
    "django-extensions>=3.2.3",
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
    "drf-yasg>=1.21.8",
]
